+++ src/interface/__init__.py
@@ -11,49 +11,16 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-bl_info = {
-    "name": "Sedaia Minecraft Utilities.",
-    "category": "Interface",
-    "author": "Sakura Sedaia <sakusedaia@outlook.com>",
-    "version": (4, 0, 0),
-    "blender": (5, 0, 0),
-    "location": "",
-    "description": "",
-    "warning": "",
-    "doc_url": "",
-    "tracker_url": "",
-    "wiki_url": "",
-}
-
-import bpy
-import bpy.utils as U
-from . import utils
-from . import prefs
-from . import interface
+from . import skin_panel
 
 modules = [
-    utils,
-    prefs,
-    interface,
-]
-
-classes = [
-
+    skin_panel,
 ]
 
 def register():
     for m in modules:
         m.register()
 
-    for c in classes:
-        U.register_class(c)
-
 def unregister():
     for m in reversed(modules):
         m.unregister()
-
-    for c in reversed(classes):
-        U.unregister_class(c)
-
-if __name__ == "__main__":
-    register()